+++ src/Concurs.java
@@ -5,7 +5,7 @@ public class Concurs extends Activitate {
 	private String disciplina;
 	private Sponsor sponsor;
 
-	public Concurs(String nume, Data dataDesfasurare, double taxa, Data termenLimitaInscriere, int nrMaximParticipanti,
+	public Concurs(String nume, Data dataDesfasurare, float taxa, Data termenLimitaInscriere, int nrMaximParticipanti,
 			int oraIncepere, String disciplina, Sponsor sponsor) {
 		super(nume, dataDesfasurare, taxa, termenLimitaInscriere, nrMaximParticipanti, oraIncepere);
 		this.disciplina = disciplina;
