+++ src/Excursie.java
@@ -4,7 +4,7 @@ public class Excursie extends Activitate {
 	
 	ArrayList<ObiectivTuristic> obiectiveTuristice;
 
-	public Excursie(String nume, Data dataDesfasurare, double taxa, Data termenLimitaInscriere, int nrMaximParticipanti,
+	public Excursie(String nume, Data dataDesfasurare, float taxa, Data termenLimitaInscriere, int nrMaximParticipanti,
 			int oraIncepere, ArrayList<ObiectivTuristic> obiectiveTuristice) {
 		super(nume, dataDesfasurare, taxa, termenLimitaInscriere, nrMaximParticipanti, oraIncepere);
 		this.obiectiveTuristice = obiectiveTuristice;
