+++ src/Sedinta.java
@@ -3,7 +3,7 @@ public class Sedinta extends Activitate {
 	private Clasa clasa;
 	private Sala sala;
 
-	public Sedinta(String nume, Data dataDesfasurare, double taxa, Data termenLimitaInscriere, int nrMaximParticipanti,
+	public Sedinta(String nume, Data dataDesfasurare, float taxa, Data termenLimitaInscriere, int nrMaximParticipanti,
 			int oraIncepere, Clasa clasa, Sala sala) {
 		super(nume, dataDesfasurare, taxa, termenLimitaInscriere, nrMaximParticipanti, oraIncepere);
 		this.clasa = clasa;
